wip: archive viewer download - Add archive-viewer.html static file - Install jszip - Add new ArchiveDownloader component - Extract all server archive code into its own module
