+++ package.json
@@ -53,6 +53,7 @@
 		"cmdk-sv": "^0.0.17",
 		"date-fns-tz": "^3.1.3",
 		"dotenv": "^16.4.5",
+		"jszip": "^3.10.1",
 		"lucide-svelte": "^0.399.0",
 		"mode-watcher": "^0.3.1",
 		"paneforge": "^0.0.5",
